Add files via upload
Java program to count total number of words in a string

This program will count total number of words in a string in Java. In this program we will read a string from the user and count total number of words in that.

The logic behind to implement this logic - Check two consecutive characters, if first character is space and next is not space, if the condition is true we will increase the counter.